routes and db setup

+++ server/routes/agents.ts
@@ -1,9 +1,9 @@
 import { Router } from 'express';
 import {
   get_all_user_agents,
-  get_user_agent_by_id,
   delete_user_agent_by_id,
   verify_user_agent_by_id,
+  is_user_agent,
 } from '../database/agents';
 import { Globals } from '../globals';
 import { Constants } from '../constants';
@@ -39,40 +39,35 @@ export function serveAgents() {
       return void res.status(500).json([]);
     }
   });
-  router.get('/agents/agent/:agent_id', async (req, res) => {
-    const { agent_id } = req.params;
-    const user_id: string = (req as any).user.id;
-    try {
-      const agent = await get_user_agent_by_id(agent_id, user_id);
-      res.json(agent);
-    } catch (error) {
-      console.error(Constants.TEXT_RED_COLOR, 'Failed to get agent:', error, Constants.TEXT_WHITE_COLOR);
-      res.status(500).json({ failure: error });
-    }
-  });
-
+  //✔️
   router.delete('/agents/agent/:agent_id', async (req, res) => {
     try {
       const { agent_id } = req.params;
       const user_id: string = (req as any).user.id;
-      await delete_user_agent_by_id(agent_id, user_id);
+
+      const isOwner = await is_user_agent(user_id, agent_id);
+      if (!isOwner) return void res.status(500).json({ success: false });
+
+      const result = await delete_user_agent_by_id(agent_id, user_id);
+
+      if (!result) return void res.status(500).json({ success: false });
 
       const socket = Globals.agentSockets.get(agent_id);
-      if (socket) {
-        socket.emit('CLOSE_AGENT', (response: any) => {
+      if (socket)
+        socket.emit('CLOSE_AGENT', () => {
           socket.disconnect();
-          return void res.status(200).json({ success: true });
         });
-      } else {
+      else {
         console.warn(Constants.TEXT_YELLOW_COLOR, 'failed getting socket for closing!');
         return void res.status(500).json({ success: false });
       }
+      return void res.status(200).json({ success: true });
     } catch (error) {
       console.error(Constants.TEXT_RED_COLOR, 'Failed to erase agent:', error, Constants.TEXT_WHITE_COLOR);
       return void res.status(500).json({ success: false });
     }
   });
-
+  //✔️
   router.get('/agents/active_status', async (req, res) => {
     try {
       const user_id: string = (req as any).user.id;
@@ -83,104 +78,93 @@ export function serveAgents() {
           status: await checkAgentStatus(agent.agent_id),
         })),
       );
-      res.status(200).json(statusUpdates);
-    } catch (err) {
-      console.error(Constants.TEXT_RED_COLOR, 'Failed to update agent statuses:', err, Constants.TEXT_WHITE_COLOR);
-      res.status(500).json({ error: 'Status update failed' });
+      return void res.status(200).json(statusUpdates);
+    } catch (error) {
+      console.error(Constants.TEXT_RED_COLOR, 'Failed to update agent statuses:', error, Constants.TEXT_WHITE_COLOR);
+      return void res.status(500).json({});
     }
   });
-
+  //✔️
   router.get('/agents/verify/:agent_id', async (req, res) => {
     try {
       const { agent_id } = req.params;
       const user_id: string = (req as any).user.id;
 
-      await verify_user_agent_by_id(agent_id, user_id);
+      const result = await verify_user_agent_by_id(agent_id, user_id);
 
-      const socket = Globals.agentSockets.get(agent_id);
-      if (socket) {
-        socket.emit('INIT_AGENT', (response: any) => {
-          if (response.status === 'initiated') {
-            res.status(200).json({ success: true });
-            return;
-          }
-        });
-      } else {
-        console.error(Constants.TEXT_RED_COLOR, 'Failed fetching socket to init!');
-        res.status(500).json({ success: false });
-      }
+      if (!result) return void res.status(500).json({ success: false });
+
+      return void res.status(200).json({ success: true });
     } catch (error) {
       console.error(Constants.TEXT_RED_COLOR, 'Failed to init:', error);
-      res.status(500).json({ success: false });
-      return;
+      return void res.status(500).json({ success: false });
     }
   });
-
+  //✔️
   router.put('/agents/monitor_status', async (req, res) => {
     try {
       const { agent_id } = req.body;
 
+      let isMonitoring = false;
+
       const socket = Globals.agentSockets.get(agent_id);
-      if (socket) {
+      if (socket)
         socket.emit('STATUS_AGENT', (response: any) => {
-          if (response.status === 'monitoring') {
-            res.status(200).json({ success: 'monitoring' });
-            return;
-          }
-          res.status(201).json({ success: 'not monitoring' });
-          return;
+          if (response.status === 'monitoring') isMonitoring = true;
         });
-      } else res.status(500).json({ success: 'not monitoring' });
+      else return void res.status(500).json({ success: 'not monitoring' });
+
+      if (isMonitoring) return void res.status(200).json({ success: 'monitoring' });
+      return void res.status(201).json({ success: 'not monitoring' });
     } catch (error) {
       console.error(Constants.TEXT_RED_COLOR, 'Failed to get agent monitoring status:', error);
-      res.status(500).json({ success: 'not monitoring' });
-      return;
+      return void res.status(500).json({ success: 'not monitoring' });
     }
   });
-
+  //✔️
   router.put('/agents/start', async (req, res) => {
     try {
       const { agent_id } = req.body;
 
+      let isRunning = false;
+
       const socket = Globals.agentSockets.get(agent_id);
-      if (socket) {
+      if (socket)
         socket.emit('START_AGENT', (response: any) => {
-          if (response.status === 'started') {
-            res.status(200).json({ success: true });
-            return;
-          }
+          if (response.status === 'started') isRunning = true;
         });
-      } else {
+      else {
         console.error(Constants.TEXT_RED_COLOR, 'Failed fetching socket to start!');
-        res.status(500).json({ success: false });
+        return void res.status(500).json({ success: false });
       }
+      if (isRunning) return void res.status(200).json({ success: true });
+      return void res.status(500).json({ success: false });
     } catch (error) {
       console.error(Constants.TEXT_RED_COLOR, 'Failed to start agent:', error);
-      res.status(500).json({ success: false });
-      return;
+      return void res.status(500).json({ success: false });
     }
   });
-
+  //✔️
   router.put('/agents/stop', async (req, res) => {
     try {
       const { agent_id } = req.body;
 
+      let isRunning = true;
+
       const socket = Globals.agentSockets.get(agent_id);
-      if (socket) {
+      if (socket)
         socket.emit('STOP_AGENT', (response: any) => {
-          if (response.status === 'stopped') {
-            res.status(200).json({ success: true });
-            return;
-          }
+          if (response.status === 'stopped') isRunning = false;
         });
-      } else {
+      else {
         console.error(Constants.TEXT_RED_COLOR, 'Failed fetching socket to stop!');
-        res.status(500).json({ success: false });
+        return void res.status(500).json({ success: false });
       }
+      if (!isRunning) return void res.status(200).json({ success: true });
+      return void res.status(500).json({ success: false });
     } catch (error) {
       console.error(Constants.TEXT_RED_COLOR, 'Failed to stop agent:', error);
-      res.status(500).json({ success: false });
-      return;
+      return void res.status(500).json({ success: false });
     }
   });
